requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.28.1
openrouter>=0.6.0
gspread==6.1.0
google-auth==2.35.0
google-genai==1.5.0
python-dotenv==1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.1
fpdf2>=2.7.0
matplotlib>=3.8.0
